+++ magda-search-api/src/main/scala/au/csiro/data61/magda/MagdaApp.scala
@@ -1,6 +1,14 @@
 package au.csiro.data61.magda
 
-import akka.actor.{Actor, ActorLogging, ActorSystem, DeadLetter, Props}
+import akka.Done
+import akka.actor.{
+  Actor,
+  ActorLogging,
+  ActorSystem,
+  CoordinatedShutdown,
+  DeadLetter,
+  Props
+}
 import akka.event.Logging
 import akka.http.scaladsl.Http
 import akka.http.scaladsl.model.HttpResponse
@@ -17,6 +25,8 @@ import au.csiro.data61.magda.search.elasticsearch.{
   DefaultClientProvider,
   ElasticSearchQueryer
 }
+
+import scala.concurrent.duration.{DurationLong}
 
 object MagdaApp extends App {
   implicit val config = AppConfig.conf()
@@ -69,7 +79,51 @@ object MagdaApp extends App {
     .orElse(Option(config.getInt("http.port")))
     .getOrElse(6102)
 
-  Http().bindAndHandle(api.routes, interface, port)
+  val bindingFuture = Http()
+    .newServerAt(interface, port)
+    .bind(api.routes)
+
+  logger.info(
+    "http.waitBeforeTermination: {}",
+    config
+      .getDuration("http.waitBeforeTermination")
+      .toMillis milliseconds
+  )
+
+  logger.info(
+    "http.hardTerminationDeadline: {}",
+    config
+      .getDuration("http.hardTerminationDeadline")
+      .toMillis milliseconds
+  )
+
+  val shutdown = CoordinatedShutdown(system)
+  shutdown.addTask(CoordinatedShutdown.PhaseServiceUnbind, "http-unbind") {
+    () =>
+      bindingFuture.flatMap(_.unbind()).map(_ => Done)
+  }
+  shutdown.addTask(
+    CoordinatedShutdown.PhaseServiceRequestsDone,
+    "http-graceful-termination"
+  ) { () =>
+    bindingFuture
+      .flatMap(
+        b =>
+          akka.pattern.after(
+            config
+              .getDuration("http.waitBeforeTermination")
+              .toMillis milliseconds,
+            system.scheduler
+          )(
+            b.terminate(
+              config
+                .getDuration("http.hardTerminationDeadline")
+                .toMillis milliseconds
+            )
+          )
+      )
+      .map(_ => Done)
+  }
 }
 
 class Listener extends Actor with ActorLogging {